@@ -22,7 +22,6 @@
     "prepare": "husky install"
   },
   "devDependencies": {
-    "@evmos-apps/config": "workspace:*",
     "@testing-library/user-event": "^14.4.3",
     "husky": "^8.0.0",
     "prettier": "^3.0.1",
@@ -34,6 +33,8 @@
     "last 3 safari version"
   ],
   "dependencies": {
+    "@babel/plugin-syntax-import-assertions": "^7.22.5",
+    "@evmos-apps/config": "workspace:*",
     "@fireworks-js/react": "^2.10.7",
     "@headlessui/react": "^1.7.15",
     "@hookform/resolvers": "^3.1.0",
@@ -41,16 +42,17 @@
     "@vitest/coverage-v8": "^0.34.1",
     "dotenv": "^16.3.1",
     "dotenv-cli": "^7.2.1",
-    "eslint": "^8.46.0",
+    "eslint": "^8.47.0",
     "eslint-interactive": "^10.8.0",
     "eslint-nibble": "^8.1.0",
     "kill-port": "^2.0.1",
+    "next": "^13.4.16",
     "nibble": "^0.0.2",
     "react-confetti-explosion": "^2.1.2",
     "react-hook-form": "^7.44.3",
     "turbo": "^1.10.12",
     "typescript": "^5.1.6",
     "vitest": "^0.34.1",
-    "zod": "^3.21.4"
+    "zod": "^3.22.1"
   }
 }
