@@ -3,30 +3,39 @@
   "version": "1.0.0",
   "description": "",
   "type": "module",
-  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^6.3.0",
-    "@typescript-eslint/parser": "^6.3.0",
-    "eslint": "^8.46.0",
-    "eslint-plugin-react": "^7.33.1"
+  "peerDependencies": {
+    "@babel/core": "^7.22.10",
+    "eslint": "^8.47.0",
+    "@babel/plugin-syntax-import-assertions": "^7.22.5",
+    "eslint-config-next": "^13.4.13",
+    "next": "^13.4.16"
   },
   "dependencies": {
+    "@babel/cli": "^7.22.10",
+    "@babel/core": "^7.22.10",
+    "@babel/eslint-parser": "^7.22.10",
+    "@babel/plugin-syntax-import-assertions": "^7.22.5",
+    "@babel/preset-env": "^7.22.10",
     "@eslint/eslintrc": "^2.1.1",
+    "@types/eslint-config-prettier": "^6.11.0",
+    "@types/minimatch": "^5.1.2",
+    "@typescript-eslint/eslint-plugin": "^6.4.0",
+    "@typescript-eslint/parser": "^6.4.0",
     "@vitejs/plugin-react": "^4.0.4",
     "eslint-config-next": "^13.4.13",
     "eslint-config-prettier": "^9.0.0",
     "eslint-config-turbo": "^1.10.12",
     "eslint-plugin-no-secrets": "^0.8.9",
+    "eslint-plugin-react": "^7.33.1",
     "eslint-plugin-sonarjs": "^0.20.0",
     "eslint-plugin-vitest": "^0.2.8",
     "jsdom": "^22.1.0",
-    "next": "^13.4.13",
     "typescript": "^5.1.6",
     "vitest": "^0.34.1"
   }
