@@ -1,13 +1,11 @@
 {
   "name": "copilot",
   "version": "1.0.3",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
+  "main": "./src/index.tsx",
+  "types": "./src/index.tsx",
   "type": "module",
   "license": "SEE LICENSE IN LICENSE",
   "scripts": {
-    "build": "tsup",
-    "dev": "tsup --watch",
     "check-types": "tsc --noEmit",
     "test": "vitest run",
     "test:watch": "vitest",
@@ -36,11 +34,11 @@
     "@types/node": "18.11.14",
     "@types/react": "18.0.26",
     "@types/react-dom": "18.0.9",
-    "eslint": "^8.46.0",
-    "ethers": "^6.7.0",
+    "eslint": "^8.47.0",
+    "ethers": "^6.7.1",
     "evmos-wallet": "workspace:*",
     "icons": "workspace:*",
-    "next": "13.4.13",
+    "next": "13.4.16",
     "react": "18.2.0",
     "react-confetti-explosion": "^2.1.2",
     "react-dom": "18.2.0",
