@@ -1,13 +1,11 @@
 {
   "name": "evmos-wallet",
   "version": "1.0.13",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
+  "main": "./src/index.tsx",
+  "types": "./src/index.tsx",
   "type": "module",
   "license": "SEE LICENSE IN LICENSE",
   "scripts": {
-    "build": "tsup",
-    "dev": "tsup --watch",
     "check-types": "tsc --noEmit",
     "test": "vitest run",
     "test:watch": "vitest",
@@ -15,9 +13,11 @@
     "lint": "eslint ."
   },
   "peerDependencies": {
-    "ethers": "^5",
+    "@tanstack/react-query": "^4.32.6",
+    "@wagmi/core": "~1.3.9",
     "react": ">=16.8.0",
-    "react-dom": ">=16.8.0"
+    "react-dom": ">=16.8.0",
+    "wagmi": "~1.3.10"
   },
   "devDependencies": {
     "@evmos-apps/config": "workspace:*",
@@ -32,6 +32,8 @@
     "@ethersproject/abstract-provider": "^5.7.0",
     "@ethersproject/bignumber": "^5.7.0",
     "@ethersproject/providers": "^5.7.0",
+    "@ethersproject/transaction": "6.0.0-beta.2",
+    "@ethersproject/transactions": "^5.7.0",
     "@evmos/address-converter": "^0.1.9",
     "@evmos/proto": "^0.1.26",
     "@evmos/provider": "^0.2.8",
@@ -40,13 +42,11 @@
     "@keplr-wallet/types": "^0.12.20",
     "@metamask/providers": "^11.1.0",
     "@reduxjs/toolkit": "^1.9.5",
-    "@tanstack/react-query": "^4.32.6",
     "@wagmi/connectors": "^2.7.0",
-    "@wagmi/core": "~1.3.9",
     "@web3modal/ethereum": "^2.7.1",
     "@web3modal/react": "^2.7.1",
     "clsx": "^2.0.0",
-    "eslint": "^8.46.0",
+    "eslint": "^8.47.0",
     "helpers": "workspace:*",
     "icons": "workspace:*",
     "immer": "9.0.21",
@@ -61,8 +61,7 @@
     "valtio": "^1.2.0",
     "viem": "^1.6.0",
     "vitest": "^0.34.1",
-    "wagmi": "^1.3.10",
-    "zod": "^3.21.4"
+    "zod": "^3.22.1"
   },
   "browserslist": [
     "last 3 chrome version",
