@@ -1,7 +1,7 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/apps/blob/main/LICENSE)
 
-module.exports = {
+export default {
   content: [
     // app content
     "src/**/*.{js,ts,jsx,tsx}",
