@@ -1,10 +1,9 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/apps/blob/main/LICENSE)
 
-const sharedConfig = require("tailwind-config/tailwind.config.js");
+const sharedConfig = require("@evmos-apps/config/tailwind/base.js");
 
 module.exports = {
-  // prefix ui lib classes to avoid conflicting with the app
   ...sharedConfig,
   plugins: [],
 };
