@@ -1,5 +1,5 @@
 import { setConfig } from "next/config";
-import config from "./next.config";
+import config from "./next.config.mjs";
 import { cleanup } from "@testing-library/react";
 import { afterEach, beforeEach, vi } from "vitest";
 import mixpanel from "mixpanel-browser";
