@@ -1,4 +1,4 @@
 import { setConfig } from "next/config";
-import config from "./next.config";
+import config from "./next.config.mjs";
 
 setConfig(config);
