@@ -1,7 +1,9 @@
+// @ts-ignore
 import { FlatCompat } from "@eslint/eslintrc";
+// @ts-ignore
+import js from "@eslint/js";
 import path from "path";
 import { fileURLToPath } from "url";
-import js from "@eslint/js";
 
 export const __filename = fileURLToPath(import.meta.url);
 
@@ -16,3 +18,14 @@ export const esCompat = new FlatCompat({
   recommendedConfig: js.configs.recommended,
   baseDirectory: __dirname,
 });
+
+export const configCompat = /** @type {(config: object) => object[]} */ (
+  esCompat.config.bind(esCompat)
+);
+
+export const extendsCompat = /** @type {(config: string) => object[]} */ (
+  esCompat.extends.bind(esCompat)
+);
+export const pluginsCompat = /** @type {(config: string) => object[]} */ (
+  esCompat.plugins.bind(esCompat)
+);
