@@ -1,3 +1,3 @@
-import sharedConfig from "@evmos-apps/config/vitest/base";
+import sharedConfig from "@evmos-apps/config/vitest/base.js";
 
 export default sharedConfig;
